finished api for the evaluation measures
