+++ carla/evaluation/__init__.py
@@ -1,8 +1,12 @@
 # flake8: noqa
 
 from .benchmark import Benchmark
-from .distance import Distance
-from .redundancy import Redundancy
-from .success_rate import SuccessRate
-from .violations import ConstraintViolation
-from .ynn import YNN
+from .catalog import (
+    YNN,
+    AvgTime,
+    ConstraintViolation,
+    Distance,
+    Redundancy,
+    SuccessRate,
+)
+from .process_nans import remove_nans
