+++ carla/evaluation/catalog/redundancy.py
@@ -3,8 +3,8 @@ from typing import List
 import numpy as np
 import pandas as pd
 
-from carla.evaluation.evaluation import Evaluation, remove_nans
-from carla.models.api import MLModel
+from carla.evaluation import remove_nans
+from carla.evaluation.api import Evaluation
 
 
 class Redundancy(Evaluation):
@@ -14,21 +14,21 @@ class Redundancy(Evaluation):
         self.columns = ["Redundancy"]
 
     def compute_redundancy(
-        self, fact: np.ndarray, cf: np.ndarray, mlmodel: MLModel, label_value: int
+        self, factual: np.ndarray, counterfactual: np.ndarray
     ) -> int:
-        red = 0
-        for col_idx in range(cf.shape[0]):  # input array has one-dimensional shape
-            if fact[col_idx] != cf[col_idx]:
-                temp_cf = np.copy(cf)
-
-                temp_cf[col_idx] = fact[col_idx]
-
-                temp_pred = np.argmax(mlmodel.predict_proba(temp_cf.reshape((1, -1))))
-
-                if temp_pred == label_value:
-                    red += 1
-
-        return red
+        redundancy = 0
+        for col_idx in range(len(counterfactual)):
+            # if feature is changed
+            if factual[col_idx] != counterfactual[col_idx]:
+                temp_cf = np.copy(counterfactual)
+                temp_cf[col_idx] = factual[col_idx]
+                # see if change is needed to flip the label
+                temp_pred = np.argmax(
+                    self.mlmodel.predict_proba(temp_cf.reshape((1, -1)))
+                )
+                if temp_pred == self.cf_label:
+                    redundancy += 1
+        return redundancy
 
     def redundancy(
         self,
@@ -36,20 +36,19 @@ class Redundancy(Evaluation):
         counterfactuals: pd.DataFrame,
     ) -> List[List[int]]:
         """
-        Computes Redundancy measure for every counterfactual
+        Computes Redundancy measure for every counterfactual.
 
         Parameters
         ----------
-        factuals: Encoded and normalized factual samples
-        counterfactuals: Encoded and normalized counterfactual samples
-        mlmodel: Black-box-model we want to discover
-        cf_label: The target counterfactual label
+        factuals:
+            Encoded and normalized factual samples.
+        counterfactuals:
+            Encoded and normalized counterfactual samples.
 
         Returns
         -------
         List with redundancy values per counterfactual sample
         """
-
         df_enc_norm_fact = factuals.reset_index(drop=True)
         df_cfs = counterfactuals.reset_index(drop=True)
 
@@ -57,8 +56,6 @@ class Redundancy(Evaluation):
             lambda x: self.compute_redundancy(
                 df_enc_norm_fact.iloc[x.name].values,
                 x.values,
-                self.mlmodel,
-                self.cf_label,
             ),
             axis=1,
         )
