+++ carla/evaluation/catalog/success_rate.py
@@ -1,0 +1,32 @@
+import pandas as pd
+
+from carla.evaluation.api.evaluation import Evaluation
+
+
+def success_rate(counterfactuals: pd.DataFrame) -> float:
+    """
+    Computes success rate for all counterfactuals.
+
+    Parameters
+    ----------
+    counterfactuals:
+        All counterfactual examples inclusive nan values.
+
+    Returns
+    -------
+
+    """
+    total_num_counterfactuals = len(counterfactuals)
+    successful_counterfactuals = len(counterfactuals.dropna())
+    success_rate = successful_counterfactuals / total_num_counterfactuals
+    return success_rate
+
+
+class SuccessRate(Evaluation):
+    def __init__(self):
+        super().__init__(None)
+        self.columns = ["Success_Rate"]
+
+    def get_evaluation(self, factuals, counterfactuals):
+        rate = success_rate(counterfactuals)
+        return pd.DataFrame([[rate]], columns=self.columns)
