+++ carla/evaluation/catalog/time.py
@@ -1,0 +1,16 @@
+import pandas as pd
+
+from carla.evaluation.api import Evaluation
+
+
+class AvgTime(Evaluation):
+    def __init__(self, hyperparameters):
+        super().__init__(hyperparameters)
+        self.time = hyperparameters["time"]
+        self.columns = ["avg_time"]
+
+    def get_evaluation(
+        self, factuals: pd.DataFrame, counterfactuals: pd.DataFrame
+    ) -> pd.DataFrame:
+        avg_time = self.time / len(counterfactuals)
+        return pd.DataFrame([[avg_time]], columns=self.columns)
