+++ carla/evaluation/catalog/ynn.py
@@ -2,30 +2,35 @@ import numpy as np
 import pandas as pd
 from sklearn.neighbors import NearestNeighbors
 
-from carla.evaluation.evaluation import Evaluation, remove_nans
+from carla.evaluation import remove_nans
+from carla.evaluation.api import Evaluation
 
 
 class YNN(Evaluation):
+    """
+    Notes
+    -----
+    - Hyperparams
+
+        * "y": int
+            Number of neighbours to use.
+        * "cf_label": int
+            What class to use as a target.
+    """
+
     def __init__(self, mlmodel, hyperparameters):
         super().__init__(mlmodel, hyperparameters)
         self.y = self.hyperparameters["y"]
         self.cf_label = self.hyperparameters["cf_label"]
-
         self.columns = ["y-Nearest-Neighbours"]
 
-    def ynn(
-        self,
-        counterfactuals,
-        factuals,
-    ):
+    def ynn(self, counterfactuals):
+
+        factuals = self.mlmodel.get_ordered_features(self.mlmodel.data.df)
+
         number_of_diff_labels = 0
-        N = counterfactuals.shape[0]
-
-        assert factuals == self.mlmodel.get_ordered_features(self.mlmodel.data.df)
         nbrs = NearestNeighbors(n_neighbors=self.y).fit(factuals.values)
-
         for i, row in counterfactuals.iterrows():
-
             if np.any(row.isna()):
                 raise ValueError(f"row {i} did not contain a valid counterfactual")
 
@@ -36,20 +41,16 @@ class YNN(Evaluation):
                 neighbour = factuals.iloc[idx]
                 neighbour = neighbour.values.reshape((1, -1))
                 neighbour_label = np.argmax(self.mlmodel.predict_proba(neighbour))
-
                 number_of_diff_labels += np.abs(self.cf_label - neighbour_label)
 
-        return 1 - (1 / (N * self.y)) * number_of_diff_labels
+        return 1 - (1 / (len(counterfactuals) * self.y)) * number_of_diff_labels
 
-    def get_evaluation(self, counterfactuals, factuals):
+    def get_evaluation(self, factuals, counterfactuals):
         counterfactuals_without_nans = remove_nans(counterfactuals)
 
         if counterfactuals_without_nans.empty:
             ynn = np.nan
         else:
-            ynn = self.ynn(
-                counterfactuals=counterfactuals_without_nans, factuals=factuals
-            )
+            ynn = self.ynn(counterfactuals=counterfactuals_without_nans)
 
-        df = pd.DataFrame([[ynn]], columns=self.columns)
-        return df
+        return pd.DataFrame([[ynn]], columns=self.columns)
